+++ Demo/Book.cs
@@ -1,0 +1,11 @@
+﻿using MongoExt.Attributes;
+using MongoExt.Interface;
+
+namespace Demo;
+
+[CachedCollection(1)]
+public class Book : IModel
+{
+	public required string Id { get; set; }
+	public required string Title { get; set; }
+}
