+++ MongoExt/Attributes/CachedCollectionAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace MongoExt.Attributes;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class CachedCollectionAttribute(int expirationMinutes) : Attribute
+{
+    public long ExpirationMinutes { get; } = expirationMinutes;
+}
