+++ MongoExt/Attributes/CollectionDefinitionAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace MongoExt.Attributes;
+
+[AttributeUsage(AttributeTargets.Interface)]
+public class CollectionDefinitionAttribute : Attribute
+{
+}
