+++ MongoExt/Attributes/CollectionImplementationAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace MongoExt.Attributes;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class CollectionImplementationAttribute : Attribute
+{
+}
