--- MongoExt/CollectionDefinitionAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace MongoDbAccess
-{
-	[AttributeUsage(AttributeTargets.Interface)]
-	public class CollectionDefinitionAttribute : Attribute
-	{
-	}
-}
