+++ MongoExt/Interface/IDbConnection.cs
@@ -1,0 +1,10 @@
+﻿using MongoDB.Driver;
+
+namespace MongoExt.Interface;
+
+public interface IDbConnection
+{
+	string GetCollectionName<T>() where T : IModel;
+
+	IMongoCollection<T> GetCollection<T>(string name) where T : IModel;
+}
