+++ MongoExt/Interface/IModel.cs
@@ -1,0 +1,6 @@
+﻿namespace MongoExt.Interface;
+
+public interface IModel
+{
+    public string Id { get; set; }
+}
