+++ MongoExt/MongoExt.csproj
@@ -4,17 +4,16 @@
     <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <Version>2.0.0</Version>
-    <Authors>FK</Authors>
-    <Company>FK Family</Company>
+    <Authors>fedark</Authors>
     <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>MongoDbAccess.snk</AssemblyOriginatorKeyFile>
+    <AssemblyOriginatorKeyFile>MongoExt.snk</AssemblyOriginatorKeyFile>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageOutputPath>d:\code\dotnet\ref\</PackageOutputPath>
+    <PackageOutputPath>package</PackageOutputPath>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="7.0.0" />
-    <PackageReference Include="MongoDB.Driver" Version="2.18.0" />
+    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="9.0.4" />
+    <PackageReference Include="MongoDB.Driver" Version="3.3.0" />
   </ItemGroup>
 
 </Project>
